requirements.txt
azure-identity==1.16.1
azure-search-documents==11.6.0b4
azure-storage-blob==12.21.0
chardet==5.2.0
environs==11.0.0
fastapi==0.111.0
gunicorn==22.0.0
loguru==0.7.2
openai==1.55.3
pdfplumber==0.11.5
python-dotenv==1.0.1
uvicorn[standard]==0.30.1
python-docx==1.1.2
